% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbegongify.R
\name{wobbegongify}
\alias{wobbegongify}
\alias{wobbegongify,SummarizedExperiment-method}
\alias{wobbegongify,SingleCellExperiment-method}
\alias{wobbegongify,DataFrame-method}
\title{Convert an object to the wobbegong format}
\usage{
wobbegongify(x, path, ...)

\S4method{wobbegongify}{SummarizedExperiment}(x, path, dense.assays = TRUE, sparse.assays = TRUE, ...)

\S4method{wobbegongify}{SingleCellExperiment}(x, path, dense.assays = TRUE, sparse.assays = TRUE, ...)

\S4method{wobbegongify}{DataFrame}(x, path, ...)
}
\arguments{
\item{x}{A supported R object, typically a SummarizedExperiment or an instance of one of its subclasses.}

\item{path}{String containing a path to the directory to dump \code{x}.}

\item{...}{Additional arguments for specific methods.}

\item{dense.assays}{Logical scalar indicating whether to save dense assays.
If \code{FALSE}, dense assays are skipped.}

\item{sparse.assays}{Logical scalar indicating whether to save dense assays.
If \code{FALSE}, sparse assays are skipped.}
}
\value{
\code{path} is populated with the contents of \code{x}.
\code{NULL} is returned invisibly.
}
\description{
Dump an object to disk in the \pkg{wobbegong} format for easy HTTP range requests.
}
\examples{
library(SingleCellExperiment)
se <- SingleCellExperiment(
    assays = list(
        counts=matrix(rpois(200, lambda=5), ncol=10), 
        logcounts=matrix(rnorm(200), ncol=10)
    ),
    colData = DataFrame(
        yy = letters[1:10],
        xx = LETTERS[1:10]
    ),
    rowData = DataFrame(row.names=sprintf("GENE_\%i", 1:20)),
    reducedDims = list(
        PCA = matrix(rnorm(50), nrow=10),
        TSNE = matrix(rnorm(20), nrow=10)
    )
)

tmp <- tempfile()
wobbegongify(se, tmp)
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
