% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wobbegongify.R
\name{wobbegongify}
\alias{wobbegongify}
\alias{wobbegongify,SummarizedExperiment-method}
\alias{wobbegongify,SingleCellExperiment-method}
\alias{wobbegongify,DataFrame-method}
\title{Convert an object to the wobbegong format}
\usage{
wobbegongify(x, path, ...)

\S4method{wobbegongify}{SummarizedExperiment}(x, path, SummarizedExperiment.assay.check = NULL, ...)

\S4method{wobbegongify}{SingleCellExperiment}(x, path, SummarizedExperiment.assay.check = NULL, ...)

\S4method{wobbegongify}{DataFrame}(x, path, ...)
}
\arguments{
\item{x}{A supported R object, typically a \link[SummarizedExperiment]{SummarizedExperiment} or an instance of one of its subclasses.}

\item{path}{String containing a path to the directory to dump \code{x}.}

\item{...}{Additional arguments for specific methods.}

\item{SummarizedExperiment.assay.check}{Function that accepts the index of the assay, the name of the assay, and the assay matrix.
It should return a logical scalar specifying whether to save the assay; if \code{FALSE}, the assay is skipped.
If \code{NULL}, no assays are skipped.}
}
\value{
\code{path} is populated with the contents of \code{x}.
\code{NULL} is returned invisibly.
}
\description{
Dump an object to disk in the \pkg{wobbegong} format for easy HTTP range requests.
}
\details{
For SummarizedExperiment objects, assays will be skipped if they are not 2-dimensional matrix-like objects of integer, logical or numeric type (according to \code{\link[DelayedArray]{type}}).
Assays will also be skipped if \code{SummarizedExperiment.assay.check} is supplied and does not return \code{TRUE} when called with the assay's details.

When passing \link[S4Vectors]{DataFrame} objects as \code{x}, columns will be skipped if they are not atomic (i.e., integer, logical, numeric or character).
Factors will be automatically converted into character vectors.
}
\examples{
library(SingleCellExperiment)
se <- SingleCellExperiment(
    assays = list(
        counts=matrix(rpois(200, lambda=5), ncol=10), 
        logcounts=matrix(rnorm(200), ncol=10)
    ),
    colData = DataFrame(
        yy = letters[1:10],
        xx = LETTERS[1:10]
    ),
    rowData = DataFrame(row.names=sprintf("GENE_\%i", 1:20)),
    reducedDims = list(
        PCA = matrix(rnorm(50), nrow=10),
        TSNE = matrix(rnorm(20), nrow=10)
    )
)

tmp <- tempfile()
wobbegongify(se, tmp)
list.files(tmp, recursive=TRUE)

}
\author{
Aaron Lun
}
